requirements.txt
numpy==1.15.0
Pillow==5.2.0
pkg-resources==0.0.0
